requirements.txt
django
gunicorn
pymupdf
pillow
pytesseract
python-docx
docx
python-bidi
arabic_reshaper
openpyxl